+++ CHANGELOG.md
@@ -13,6 +13,7 @@ incremented upon a breaking change and the patch version will be incremented for
 **Added**
 
 - add more stake program related methods ([441](https://github.com/Ackee-Blockchain/trident/pull/441))
+- add remove functionality to AddressStorage ([442](https://github.com/Ackee-Blockchain/trident/pull/442))
 
 **Removed**